A Complete Example Using a Functional Component
It took me a while to figure out how to properly write the code in a functional component format, rather than a class. So, I am suggesting adding a complete example using a functional component. Moreover, the example provides a complete, working example, which also was not clear for me from the current readme file.